change float to int on frontend, changed all decimals to int on backend

+++ Trail Composer/Trail Composer/Controllers/POIController.cs
@@ -43,8 +43,8 @@ namespace Trail_Composer.Controllers
 
         [Authorize]
         [HttpGet("list/filtered")]
-        public async Task<IActionResult> GetFilteredPOIList([FromQuery] int[] countryIds, [FromQuery] decimal minLatitude, [FromQuery] decimal maxLatitude,
-            [FromQuery] decimal minLongitude, [FromQuery] decimal maxLongitude)
+        public async Task<IActionResult> GetFilteredPOIList([FromQuery] int[] countryIds, [FromQuery] int minLatitude, [FromQuery] int maxLatitude,
+            [FromQuery] int minLongitude, [FromQuery] int maxLongitude)
         {
             var userId = TCUserDTO.GetUserIdFromContext(this.HttpContext);
             var result = await _poiService.GetFilteredPoiListAsync(userId ,countryIds, minLatitude, maxLatitude, minLongitude, maxLongitude);
@@ -53,8 +53,8 @@ namespace Trail_Composer.Controllers
 
         [Authorize]
         [HttpGet("list/user/filtered")]
-        public async Task<IActionResult> GetFilteredUserPOIList([FromQuery]int[] countryIds, [FromQuery] decimal minLatitude, [FromQuery] decimal maxLatitude,
-            [FromQuery] decimal minLongitude, [FromQuery] decimal maxLongitude)
+        public async Task<IActionResult> GetFilteredUserPOIList([FromQuery]int[] countryIds, [FromQuery] int minLatitude, [FromQuery] int maxLatitude,
+            [FromQuery] int minLongitude, [FromQuery] int maxLongitude)
         {
             var userId = TCUserDTO.GetUserIdFromContext(this.HttpContext);
             var result = await _poiService.GetFilteredUserPoiListAsync(userId, countryIds, minLatitude, maxLatitude, minLongitude, maxLongitude);
